+++ admin/activeplanet.php
@@ -11,7 +11,6 @@ restrictAccess($user, LEVEL_OPERATOR);
 includeLang('admin');
 
 $parse          = $lang;
-$parse['dpath'] = $dpath;
 
 $PageTPL        = gettemplate('admin/activeplanet_body');
 $AllActivPlanet = doquery("SELECT * FROM {{table}} WHERE `last_update` >= '". (time()-15 * 60) ."' ORDER BY `id` ASC", 'planets');